move scroll functionality for main banner to Main.tsx
